requirements.txt
ctransformers==0.2.27
sentence-transformers==3.0.1
pinecone-client
langchain==0.2.6
flask
langchain-community
langchain-huggingface
pypdf
langchain-pinecone
python-dotenv
-e .